Reservation

+++ project_4_web_service/src/main/java/io/ab/library/repository/AccountRepository.java
@@ -8,7 +8,7 @@ import io.ab.library.model.Account;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface AccountRepository extends CrudRepository<Account, Long> {
+public interface AccountRepository extends CrudRepository<Account, Integer> {
 	
 	Account findByEmailIs(String email);
